removed MainnetView; fixed styles
