+++ src/renderer/views/wallet/CreateView/CreateView.styles.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const Container = styled('div')`
+  flex: 1;
+`
